+++ features/quiz/index.ts
@@ -1,0 +1,3 @@
+export { makeQuizList } from './model/makeQuizList'
+export { default as AnswerItem } from './ui/AnswerItem'
+export { default as QuizItem } from './ui/QuizItem'
