+++ widgets/quiz-container/index.ts
@@ -1,2 +1,2 @@
-export { default as QuizItem } from './ui/QuizItem'
-export { default as AnswerItem } from './ui/AnswerItem'
+export { default as QuizItem } from '../../features/quiz/ui/QuizItem'
+export { default as AnswerItem } from '../../features/quiz/ui/AnswerItem'
